+++ core/src/main/java/arez/annotations/Track.java
@@ -87,5 +87,5 @@ public @interface Track
    *
    * @return true if the observer can invoke actions, false otherwise.
    */
-  boolean canNestActions() default false;
+  boolean nestedActionsAllowed() default false;
 }