+++ processor/src/test/resources/expected/com/example/autorun/Arez_NestedActionsAllowedAutorunModel.java
@@ -16,7 +16,7 @@ import javax.annotation.Nullable;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-public final class Arez_CanNestActionsAutorunModel extends CanNestActionsAutorunModel implements Disposable, Identifiable<Integer>, DisposeTrackable {
+public final class Arez_NestedActionsAllowedAutorunModel extends NestedActionsAllowedAutorunModel implements Disposable, Identifiable<Integer>, DisposeTrackable {
   private static volatile int $$arezi$$_nextId;
 
   private final int $$arezi$$_id;
@@ -33,14 +33,14 @@ public final class Arez_CanNestActionsAutorunModel extends CanNestActionsAutorun
   @Nonnull
   private final Observer $$arez$$_doStuff;
 
-  public Arez_CanNestActionsAutorunModel() {
+  public Arez_NestedActionsAllowedAutorunModel() {
     super();
     this.$$arezi$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
     this.$$arezi$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? $$arezi$$_nextId++ : 0;
     if ( Arez.shouldCheckApiInvariants() ) {
       this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
     }
-    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().component( "CanNestActionsAutorunModel", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> $$arezi$$_preDispose() ) : null;
+    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().component( "NestedActionsAllowedAutorunModel", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> $$arezi$$_preDispose() ) : null;
     this.$$arezi$$_disposeNotifier = new DisposeNotifier();
     this.$$arez$$_doStuff = $$arezi$$_context().observer( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".doStuff" : null, () -> super.doStuff(), Flags.RUN_LATER | Flags.NESTED_ACTIONS_ALLOWED );
     if ( Arez.shouldCheckApiInvariants() ) {
@@ -60,7 +60,7 @@ public final class Arez_CanNestActionsAutorunModel extends CanNestActionsAutorun
 
   final ArezContext $$arezi$$_context() {
     if ( Arez.shouldCheckApiInvariants() ) {
-      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_context' invoked on uninitialized component of type 'CanNestActionsAutorunModel'" );
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_context' invoked on uninitialized component of type 'NestedActionsAllowedAutorunModel'" );
     }
     return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
   }
@@ -80,9 +80,9 @@ public final class Arez_CanNestActionsAutorunModel extends CanNestActionsAutorun
 
   String $$arezi$$_name() {
     if ( Arez.shouldCheckApiInvariants() ) {
-      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_name' invoked on uninitialized component of type 'CanNestActionsAutorunModel'" );
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_name' invoked on uninitialized component of type 'NestedActionsAllowedAutorunModel'" );
     }
-    return "CanNestActionsAutorunModel." + $$arezi$$_id();
+    return "NestedActionsAllowedAutorunModel." + $$arezi$$_id();
   }
 
   private void $$arezi$$_preDispose() {
@@ -140,10 +140,10 @@ public final class Arez_CanNestActionsAutorunModel extends CanNestActionsAutorun
     if ( Arez.areNativeComponentsEnabled() ) {
       if ( this == o ) {
         return true;
-      } else if ( null == o || !(o instanceof Arez_CanNestActionsAutorunModel) ) {
+      } else if ( null == o || !(o instanceof Arez_NestedActionsAllowedAutorunModel) ) {
         return false;
       } else {
-        final Arez_CanNestActionsAutorunModel that = (Arez_CanNestActionsAutorunModel) o;
+        final Arez_NestedActionsAllowedAutorunModel that = (Arez_NestedActionsAllowedAutorunModel) o;
         return $$arezi$$_id() == that.$$arezi$$_id();
       }
     } else {